+++ js/inputmask.extensions.js
@@ -9,9 +9,9 @@ Optional extensions on the jquery.inputmask base
 */
 (function(factory) {
 		if (typeof define === "function" && define.amd) {
-			define(["jquery", "inputmask"], factory);
+			define(["inputmask.dependencyLib", "inputmask"], factory);
 		} else if (typeof exports === "object") {
-			module.exports = factory(require("jquery"), require("./inputmask"));
+			module.exports = factory(require("./inputmask.dependencyLib.jquery"), require("./inputmask"));
 		} else {
 			factory(jQuery, window.Inputmask);
 		}